Eureka

+++ clientSuiviMarches/src/main/java/ma/ensak/clientSuiviMarches/controller/ServiceController.java
@@ -28,7 +28,7 @@ public class ServiceController {
 	@RequestMapping(value="addOrUpdateServiceRedirect")
 	public String addOrUpdateService(HttpServletRequest request , Model model ,@RequestParam(defaultValue="")String message ,@RequestParam(defaultValue="") String idd  , Service service)
 	{
-		if(!Roles.isALL(request)) return "redirect:/authentification"; 
+		if(!Roles.isALL(request) || Roles.isChefService(request) || Roles.isFonctionnaire(request)) return "redirect:/authentification"; 
 		
 		model.addAttribute("message", "".equals(message) ? "" : message );
 		model.addAttribute("Sectiontitle", "Services");
@@ -40,7 +40,7 @@ public class ServiceController {
 	public ModelAndView addOrUpdateService( HttpServletRequest request  ,RedirectAttributes redirectAttributes, Service service)
 	{
 		
-		if(!Roles.isALL(request)) return new ModelAndView("redirect:/authentification");
+		if(!Roles.isALL(request) || Roles.isChefService(request) || Roles.isFonctionnaire(request)) return new ModelAndView("redirect:/authentification");
 		
 		Long id = service.getId();
 		ModelAndView modelAndView = null ;
@@ -60,7 +60,7 @@ public class ServiceController {
 	@RequestMapping(value="deleteService")
 	public ModelAndView deleteService( HttpServletRequest request , @RequestParam(defaultValue="")String idd)
 	{
-		if(!Roles.isALL(request)) return new ModelAndView("redirect:/authentification");
+		if(!Roles.isALL(request) || Roles.isChefService(request) || Roles.isFonctionnaire(request)) return new ModelAndView("redirect:/authentification");
 		
 		microSEProxy.deleteService(Long.parseLong(idd));;
 		return new ModelAndView("redirect:/getServices","message","Suppréssion réussie");
@@ -69,7 +69,7 @@ public class ServiceController {
 	@RequestMapping(value="getServices")
 	public String getServices( HttpServletRequest request , Model model,@RequestParam(defaultValue="") String message)
 	{
-		if(!Roles.isALL(request)) return "redirect:/authentification";
+		if(!Roles.isALL(request) || Roles.isChefService(request) || Roles.isFonctionnaire(request)) return "redirect:/authentification";
 		
 		model.addAttribute("message", "".equals(message) ? "" : message );
 		model.addAttribute("services", microSEProxy.getAllServices());
